requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1
markdown==3.5.1
tinydb==4.8.0